# openapi.yaml
openapi: 3.0.3
info:
  title: Unikol API
  description: >
    API REST de la plateforme SaaS de gestion scolaire Unikol.
    Toutes les routes (sauf /auth/*) exigent un JWT Bearer valide.
    Le tenant (SchoolId) est dérivé du claim `schoolId` du token, jamais d'un paramètre de requête.
    Référence fonctionnelle complète : docs/Volume_4_API_Design.md
  version: "1.1.0"
servers:
  - url: https://api.sama-ecole.sn/v1
    description: Production
  - url: https://staging-api.sama-ecole.sn/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Développement local
security:
  - bearerAuth: []
...
paths:
...
  /enrollments:
    post:
      tags: [Enrollments]
      summary: Créer une inscription/réinscription (déclenche automatiquement le calcul Finance)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollmentCreateRequest' }
      responses:
        '201':
          description: Inscription créée, frais calculés automatiquement (renvoie le reçu)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollmentReceipt' }
        '403': { description: Rôle non autorisé (le service Finance ne compose pas de montant dû) }
        '409': { description: Élève déjà inscrit pour l'année active }
        '422': { description: Aucune année active, ou données invalides }
  /enrollments/{id}:
    delete:
      tags: [Enrollments]
      summary: Annuler une inscription saisie par erreur (jamais de suppression physique)
      description: >-
        Réservé au Directeur et au Secrétariat. Refusée en 409 si un paiement a déjà été encaissé —
        utiliser POST /enrollments/{id}/status pour un abandon ou un transfert à la place.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: rowVersion
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Inscription annulée (statut Cancelled) }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '404': { description: Inscription introuvable dans l'établissement courant }
        '409': { description: 'Jeton de concurrence périmé, déjà annulée, ou un paiement existe déjà' }
  /enrollments/{id}/status:
    post:
      tags: [Enrollments]
      summary: Déclarer un abandon ou un transfert en cours d'année
      description: >-
        Réservé au Directeur et au Secrétariat. L'inscription n'est PAS annulée : elle change de
        statut. L'élève sort des futures listes de présence actives ; ses notes et paiements déjà
        enregistrés restent intacts.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeEnrollmentStatusRequest' }
      responses:
        '204': { description: Statut mis à jour }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '404': { description: Inscription introuvable dans l'établissement courant }
        '409': { description: 'Jeton de concurrence périmé, ou l''inscription est déjà dans un statut terminal' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /enrollments/{id}/receipt:
    get:
      tags: [Enrollments]
      summary: Reçu d'inscription (ticket JGK-E01 — JSON ; le PDF pixel-fidèle est JGK-E02)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Reçu (détail des frais figé à l'inscription)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollmentReceipt' }
        '404': { description: Inscription introuvable dans l'établissement courant }
  /enrollments/{id}/receipt/pdf:
    get:
      tags: [Enrollments]
      summary: Reçu d'inscription officiel en PDF (ticket JGK-E02 — reproduit docs/design-references/receipt-reference.png)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Le reçu au format PDF (téléchargement)
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '404': { description: Inscription introuvable dans l'établissement courant }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Format normalisé — voir docs/Volume_4_API_Design.md §0.4
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
        traceId: { type: string }
...
    EnrollmentCreateRequest:
      type: object
      description: >-
        Deux modes selon `type`. NewEnrollment : l'élève est CRÉÉ dans le flux (fullName, birthDate,
        gender requis) et son matricule généré. ReEnrollment : `studentId` d'un élève existant.
        Ni `schoolYearId` ni `totalDue` : l'année ACTIVE est résolue serveur, et le montant est
        CALCULÉ à partir du barème de la classe — jamais fourni par le client (règles #4 et #10).
      required: [type, classroomId]
      properties:
        type: { type: string, enum: [NewEnrollment, ReEnrollment] }
        classroomId: { type: string, format: uuid }
        studentId:
          type: string
          format: uuid
          description: Requis pour ReEnrollment ; ignoré pour NewEnrollment.
        fullName: { type: string, maxLength: 200, description: NewEnrollment uniquement }
        birthDate: { type: string, format: date, description: NewEnrollment uniquement }
        gender: { type: string, enum: [M, F], description: NewEnrollment uniquement }
        guardianName: { type: string }
        guardianPhone: { type: string }
        collectedFees:
          type: array
          description: >-
            Frais RÉELLEMENT réglés au guichet le jour de l'inscription. Absent ou vide : dossier
            ouvert sans versement — aucun paiement n'est créé et le reçu s'imprime à 0. Le client
            désigne QUOI est réglé, jamais COMBIEN : les montants sont repris du barème que le serveur
            vient de figer (règle #4). Une catégorie ne peut apparaître qu'une fois. Un versement est
            matérialisé par un Payment portant le MÊME numéro de reçu que l'inscription.
          items:
            type: object
            required: [feeCategoryId]
            properties:
              feeCategoryId: { type: string, format: uuid }
              months:
                type: integer
                default: 1
                minimum: 1
                maximum: 24
                description: >-
                  Mois réglés d'avance sur une mensualité (borné serveur au nombre de mois facturés).
                  Ignoré sur un frais ponctuel : il se règle en entier ou pas du tout.
        paymentMethod:
          type: string
          enum: [Cash, Cheque, Transfer, MobileMoney]
          default: Cash
          description: Mode de règlement du versement du jour ; ignoré si `collectedFees` est vide.
    EnrollmentReceipt:
      type: object
      description: >-
        Reçu d'inscription (JGK-E01), imprimé en A5 paysage. Le détail des frais est figé à
        l'inscription. Deux montants à ne pas confondre : `totalCollected` est ce qui est RÉELLEMENT
        entré en caisse ce jour-là (seul total imprimé en gras, ventilé par `collectedLines`), tandis
        que `totalDue` est le dû annuel, rappelé sous le tableau avec le reste à payer.
      properties:
        enrollmentId: { type: string, format: uuid }
        receiptNumber: { type: string, description: 'Numéro officiel unique du reçu (JGK-E02), ex. « REC-2025-0002 ». Porté aussi par le versement du jour.' }
        schoolName: { type: string }
        schoolAddress: { type: string, nullable: true }
        schoolPhone: { type: string, nullable: true }
        schoolEmail: { type: string, nullable: true }
        schoolNinea: { type: string, nullable: true, description: 'Mention légale de l''en-tête ; absente si non saisie.' }
        schoolRegistreCommerce: { type: string, nullable: true, description: 'RCCM — idem.' }
        schoolCity: { type: string, nullable: true, description: 'Ville du bas de reçu (« Fait à … »), dérivée de l''adresse de l''école.' }
        matricule: { type: string }
        studentFullName: { type: string }
        classroomName: { type: string }
        classroomLevel: { type: string }
        schoolYearLabel: { type: string }
        guardianName: { type: string, nullable: true }
        guardianPhone: { type: string, nullable: true }
        type: { type: string, enum: [NewEnrollment, ReEnrollment] }
        status: { type: string, enum: [Pending, Confirmed, Cancelled, DroppedOut, Transferred] }
        enrolledAt: { type: string, format: date-time }
        totalDue: { type: number, description: 'Dû ANNUEL figé à l''inscription — jamais le montant du reçu.' }
        lines:
          type: array
          description: Détail du dû annuel, figé à l'inscription.
          items:
            type: object
            properties:
              designation: { type: string }
              isRecurring: { type: boolean }
              unitAmount: { type: number }
              months: { type: integer }
              lineTotal: { type: number }
        collectedLines:
          type: array
          description: >-
            Ventilation de l'encaissement du jour — une ligne par frais réellement réglé au guichet.
            Vide si le dossier a été ouvert sans versement.
          items:
            type: object
            properties:
              designation: { type: string }
              isRecurring: { type: boolean }
              months: { type: integer, description: 'Mois couverts pour une mensualité ; 1 pour un frais ponctuel.' }
              amount: { type: number }
        totalCollected: { type: number, description: 'Somme réellement encaissée ce jour — « TOTAL ENCAISSÉ » du reçu.' }
        remainingBalance: { type: number, description: 'totalDue − totalCollected, rappelé sous le tableau.' }
        paymentMethod:
          type: string
          nullable: true
          enum: [Cash, Cheque, Transfer, MobileMoney]
          description: 'Mode de règlement du versement du jour ; null quand rien n''a été encaissé.'
    ChangeEnrollmentStatusRequest:
      type: object
      required: [newStatus, rowVersion]
      description: >-
        Déclare un abandon ou un transfert en cours d'année (voir POST /enrollments/{id}/status).
        Seuls DroppedOut et Transferred sont acceptés ici — Cancelled passe par
        DELETE /enrollments/{id}, réservé aux inscriptions jamais encaissées.
      properties:
        newStatus: { type: string, enum: [DroppedOut, Transferred] }
        rowVersion: { type: integer, format: int64 }
...